fix items switcher not being full generic
